pyproject.toml
[tool.poetry]
name = "repl_python3_SpotifyPartyController"
version = "0.1.0"
description = ""
authors = ["DillonB07 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
